+++ src/components/HomePage/HomePage.tsx
@@ -5,17 +5,16 @@ import { SliderSwiper } from '../SliderSwiper/SliderSwiper';
 import { ShopByCategory } from '../ShopByCategory/ShopByCategory';
 import { ProductSlider } from '../ProductCard/ProductCard';
 import { useEffect, useState } from 'react';
-import { Product } from '../../types/typeRpoduct';
-import { fetchProducts } from '../../utils/api';
+import { Product } from '../../types/ProductDetails';
+import { fetchAllProducts } from '../../utils/api';
 import { Loader } from '../Loader/Loader';
 import { NameSlider } from '../../nameslider';
-
 export const HomePage = () => {
   const [phones, setPhones] = useState<Product[]>([]);
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
-    fetchProducts()
+    fetchAllProducts()
       .then(data => {
         setPhones(data.filter(dat => dat.category === 'phones'));
       })
@@ -28,20 +27,16 @@ export const HomePage = () => {
     <main className="main__homepage">
       <h1 className="homepage__title">Welcome to Nice Gadgets store!</h1>
       {loading && <Loader />}
-
       {!loading && (
         <>
           <SliderSwiper />
-
           <div className="homepage__product">
             <ProductSlider products={phones} title={NameSlider.Brand} />
           </div>
-
           <div className="homepage__category">
             <h2 className="category">Shop by category</h2>
             <ShopByCategory />
           </div>
-
           <div className="homepage__Hotproduct">
             <ProductSlider
               products={phones}
