+++ src/types/ProductDetails.tsx
@@ -34,3 +34,18 @@ export interface Tablets extends ProductDetails {
   camera: string;
   zoom: string;
 }
+
+export interface Product {
+  id: number;
+  category: string;
+  itemId: string;
+  name: string;
+  fullPrice: number;
+  price: number;
+  screen: string;
+  capacity: string;
+  color: string;
+  ram: string;
+  year: number;
+  image: string;
+}
