+++ src/types/typeRpoduct.tsx
@@ -1,14 +1,14 @@
-export interface Product {
-  id: number;
-  category: string;
-  itemId: string;
-  name: string;
-  fullPrice: number;
-  price: number;
-  screen: string;
-  capacity: string;
-  color: string;
-  ram: string;
-  year: number;
-  image: string;
-}
+// export interface Product {
+//   id: number;
+//   category: string;
+//   itemId: string;
+//   name: string;
+//   fullPrice: number;
+//   price: number;
+//   screen: string;
+//   capacity: string;
+//   color: string;
+//   ram: string;
+//   year: number;
+//   image: string;
+// }
